exception handling, test

+++ src/main/java/com/example/mywebquizengine/service/PhotoService.java
@@ -1,7 +1,7 @@
 package com.example.mywebquizengine.service;
 
 import com.example.mywebquizengine.model.User;
-import com.example.mywebquizengine.model.exception.EmptyFileException;
+import com.example.mywebquizengine.model.exception.LogicException;
 import com.example.mywebquizengine.model.userinfo.Photo;
 import com.example.mywebquizengine.repos.PhotoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -52,9 +52,8 @@ public class PhotoService {
 
     }
 
-
     @Transactional
-    public String uploadPhoto(MultipartFile file, String username) throws EmptyFileException, IOException {
+    public String uploadPhoto(MultipartFile file, String username) throws LogicException, IOException {
         if (!file.isEmpty()) {
             try {
                 String uuid = UUID.randomUUID().toString();
@@ -81,15 +80,17 @@ public class PhotoService {
             } catch (IOException e) {
                 throw new IOException("Input-Output error");
             }
-        } else throw new EmptyFileException("File is empty");
+        } else throw new LogicException("File is empty");
     }
 
-    public void deletePhoto(Long photoId, String authUsername) throws IllegalAccessError, EntityNotFoundException {
+    public void deletePhoto(Long photoId, String authUsername) throws LogicException {
         Optional<Photo> optionalPhoto = photoRepository.findById(photoId);
         if (optionalPhoto.isPresent()) {
             String photoLoaderUsername = photoRepository.getPhotoLoaderUsername(photoId);
             if (photoLoaderUsername.equals(authUsername)) {
-                photoRepository.deleteById(photoId);
+                if (photoRepository.getPhotoCountByUsername(authUsername) > 1) {
+                    photoRepository.deleteById(photoId);
+                } else throw new LogicException("You must have at least one photo");
             } else throw new SecurityException("You are not loader of this photo");
         } else throw new EntityNotFoundException("Photo with given photoId not found");
     }
